+++ docs/concepts/other_techniques.md
@@ -27,8 +27,8 @@ input_tokens_required = client.count_tokens(
 if input_tokens_required > input_token_limit:
     # In this case we know that an API call with the current content will exceed the input token limit for the current model.
 ```
-
 - Check whether the output token limit has been exceeded. It is important to note that we can not predict this before making the API call, and intead have to make the API call and then check whether it succeeded. This is discussed more in the [Error Handling section](https://phil-daniel.github.io/gemini-batcher/concepts/error_handling.html#max_tokens-finish-reason) of this guide.
+
 ```python
 response = client.models.generate_content(
     model="gemini-2.5-flash",
@@ -91,21 +91,83 @@ while len(queue) > 0:
     # Add logic for handling valid responses here...
 ```
 
-TODO: Add visualisation of this process
-
-
 ## Semantic Chunking and Batching
 
 ### Semantic Chunking
 
-Up until now, all of the chunking techniques discussed have broken up the content purely based on its length rather than considering the meaning and structure of the content itself.
+Up until now, all of the chunking techniques discussed have broken up the content purely based on its length, rather than considering the meaning and structure of the content itself. Semantic chunking takes a different approach. Instead of breaking up the text arbitrarily, content is split at meaningful boundaries, such as topic shifts. This increases the likelihood that the answer to a question is entirely contained within a single chunk, which can result in improved responses.
 
-Semantic chunking takes a different approach, instead of breaking up the text arbitrarily, content is split at meaningful boundaries, such as topic shifts.
+TO perform semantic chunking, we begin by splitting the transcript into individual sentences, and then converting these sentences into embeddings (high dimensional vectors), similar to the steps required for LLMs. We can then calculate the consine similarity between adjacent sentence embeddings, where a large change indicates a potential boundary position for chunking.
 
-TODO: Add code sample
+```python
+from sentence_transformers import SentenceTransformer
+from sklearn.metrics.pairwise import cosine_similarity
+
+# Splitting sentences and stripping excess detail
+sentences = re.split(r'(?<=[.!?])\s+', content)
+sentences = [sentence.strip() for sentence in sentences]
+
+model = SentenceTransformer(transformer_model)
+
+# Creating sentence embeddings using the SentenceTransformer model
+sentence_embeddings = model.encode(sentences)
+
+# Calculating the similarity between adjacent embeddings
+similarities = []
+for i in range(len(sentence_embeddings) - 1):
+    # Reshape for cosine_similarity: (1, n_features) for each vector
+    s1 = sentence_embeddings[i].reshape(1, -1)
+    s2 = sentence_embeddings[i+1].reshape(1, -1)
+    similarity = cosine_similarity(s1, s2)[0][0]
+    similarities.append(similarity)
+
+# Calculating a threashold value for cosine similarity.
+mean = np.mean(similarities)
+std_dev = np.std(similarities)
+similarity_threashold = mean - (std_dev * threashold_factor)
+
+boundaries = [0]
+current_chunk_start_pos = 0
+for i in range(len(similarities)):
+    # Checking if there is a natural boundary.
+    if similarities[i] < similarity_threashold and (i + 1) - current_chunk_start_pos >= min_sentences_per_chunk:
+        boundaries.append(i+1)
+        current_chunk_start_pos = i + 1
+    elif (i+1) - current_chunk_start_pos >= max_sentences_per_chunk:
+        boundaries.append(i+1)
+        current_chunk_start_pos = i + 1
+        
+# Adding the end point if it has not already been added
+if boundaries[-1] != len(similarities) + 1:
+    boundaries.append(len(similarities) + 1)
+
+# Creating the chunks based on the boundaries.
+content_chunks = []
+for i in range(len(boundaries) - 1):
+    content_chunks.append(" ".join(sentences[boundaries[i] : boundaries[i+1]]))
+```
 
 ### Semantic Batching
 
-We can continue with the semantic approach by also batching questions based on their meanings. There are two main approaches to this, questions can be batched together based on their similarity to one another, or questions can be batched to chunks based on their similarity to the chunk's contents. The code required to complete both of these techniques is very similar to that shown in the semantic chunking.
+We can continue with the semantic approach by also batching questions based on their meanings. There are two main approaches to this:
+1. Questions can be batched together based on their similarity to one another.
+2. Questions can be batched to chunks based on their similarity the chunk's contents. The advantage of this technique is that the model doesn't need to be queried with irrelvant question-chunk pairs, further improving the efficiency and token usage.
 
-TODO: Add code
+```python
+# Creating a batch for each chunk. Each batch only contains the questions for its respective chunks.
+question_batches = [[] for _ in range(len(chunked_content))]
+
+# Creating embeddings for each question.
+question_embeddings = model.encode(questions)
+# Creating an embeddings for each chunk - not each sentence in a chunk.
+chunk_embeddings = model.encode(chunked_content)
+
+for i in range(len(question_embeddings)):
+    # Calculating the similarity to each chunk.
+    chunk_similarity = cosine_similarity(question_embeddings[i].reshape(1, -1), chunk_embeddings)[0]
+    # Finding the most similar chunk and adding the question to its batch.
+    most_similar_chunk = np.argmax(chunk_similarity)
+    question_batches[most_similar_chunk].append(questions[i])
+```
+
+It is worth noting that while semantic chunking and batching can improve performance and reduce token usage, it does require additional computation (such as calculating embeddings) that the fixed techniques do not, meaning that the best approach will depend on your specific use case and constraints.
