--- gemini_batcher/input_handler/otherinputs.py
@@ -1,25 +1,0 @@
-class VideoFileInput():
-
-    def __init__(
-        self,
-        filepath : str
-    ):
-        self.filepath = filepath
-        return
-    
-class AudioFileInput():
-
-    def __init(
-        self,
-        filepath : str
-    ):
-        pass
-
-class YoutubeInput():
-
-    def __init__(
-        self,
-        link : str
-    ):
-        # TODO: Error checking that link exists.
-        pass
